Rook and Bishop Move Generation Added.

+++ src/Chess/Moves.cc
@@ -2,11 +2,11 @@
 #include "../engine.hh"
 
 uint64 MoveGenerator::RookAttacks(Square square, uint64 occupancy){
-    Magic& magic = RMagics[(int)square];
+    Magic& magic = RMagics[square];
     return magic.AttackTable[magic.GetIndexFromOccupancy(occupancy)];
 }
 
 uint64 MoveGenerator::BishopAttacks(Square square, uint64 occupancy){
-    Magic& magic = BMagics[(int)square];
+    Magic& magic = BMagics[square];
     return magic.AttackTable[magic.GetIndexFromOccupancy(occupancy)];
 }
